week1 & week2 submitted. (#30)
* Add files via upload

* Add files via upload

* week2 solution; RBF and Child tx.

The script only assumes that Bitcoin Core is installed and can be accessed through bitcoind and bitcoin-cli.
It checks if bitcoind is running, and starts the node if it is not.

Every other case is handled by the script.

* Delete shreyan_week2.sh

* shreyan_week2 solution;RBF and Child tx

The script checks if bitcoind is already running. If not, it starts the node.
The script assumes Bitcoin Core is already installed and can be accessed locally using bitcoind and bitcoin-cli.
All other cases shall be handled by the script.

* Delete shreyan.sh

Functions not well-defined and script not machine agnostic.
The solution is in file shreyan_week1.sh. Kindly, check.
Apologies for deleting files. Learning git and github.